Addressing #8.
Here's the process (this needs to be added to the README of blog-posts): Add your images to the top-level images directory of the blog-posts repository and add a thumbnail property, e.g.:

```
thumbnail: images/<your_file_name>
```

The main super irritating thing that I haven't automated is that we need to create softlinks from this repo's `images` directory to `blog-post`'s image directory. I tried some stuff with prefixes but I couldn't grab them in liquid. It must be possible, but because (a) they live in an "ignored" directory (`_posts/...`) and (b) the blog posts are generated files, I would need to do more investigation to figure out what the path should actually be. In the meantime, it just means we need to do an extra step before merging from `blog` into `master`.

On a related note, I would like to alternate where the image appears -- as you can see with this commit, the images are all on the left, which makes the page feel imblanaced. However, I am not sure how to have local variables in Liquid or whether we can grab the index and choose the side with a mod.